requirements.txt
requests
beautifulsoup4
pandas
openpyxl
cohere
selenium
webdriver-manager
serpapi
python-dateutil
